\name{SAMM-package}
\alias{SAMM-package}
\alias{SAMM-package}
\docType{package}
\title{
\packageTitle{SAMM}
}
\description{
\packageDescription{SAMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SAMM}
\packageIndices{SAMM}
 This program can be used to fit Gaussian linear mixed models (LMM). Univariate and multivariate response models, multiple variance components, as well as, certain correlation and covariance structures are supported. In many occations, the user can pick one of the several mixed model fitting algorithms, which are explained further in the details section. Some algorithms are specific to certain types of models (univariate or multivariate, diagonal or nondiagonal residual, one or multiple variance components, etc,...).

}
\author{
\packageAuthor{SAMM}

Maintainer: \packageMaintainer{SAMM}
}
\references{
Bates, Douglas M. "lme4: Mixed-effects modeling with R." URL http://lme4. r-forge. r-project. org/book (2010).

Zhou, Hua, et al. "MM Algorithms for Variance Components Models." arXiv preprint arXiv:1509.07426 (2015).

Zhou, Xiang, and Matthew Stephens. "Efficient algorithms for multivariate linear mixed models in genome-wide association studies." Nature methods 11.4 (2014): 407.

Kang, Hyun Min, et al. "Efficient control of population structure in model organism association mapping." Genetics 178.3 (2008): 1709-1723.

Gilmour, Arthur R., Robin Thompson, and Brian R. Cullis. "Average information REML: an efficient algorithm for variance parameter estimation in linear mixed models." Biometrics (1995): 1440-1450.


}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
x=2
}
